+++ Changelog.md
@@ -11,6 +11,7 @@
 - Fix: The One Probe doesn't display some of SecurityCraft's blocks correctly
 - Fix: Potential startup crash in combination with other mods
 - Fix: Admin Tool description in JEI is incorrect
+- Fix: Crash when inserting a module into a single Inventory Scanner
 
 --------------------------Changelog for v1.9.4 of SecurityCraft--------------------------